Add progress bar for chain sync
This commit address github issue #10:
https://github.com/lightninglabs/lightning-app/issues/10

This update uses (the currently proposed) syncedHeight and blockHeight
from the gRPC interface to lnd to determine the percentage of
completion toward the chain sync.

While the chain is not fully synced, the Sidebar makes a request to
update the account every 1000ms so the progress display is reasonably
smooth without spamming the lnd. The progress is displayed as an
absolutely positioned div over the status div using a complementary
color.